big commit

+++ code/04b_geo_rain.R
@@ -1,6 +1,6 @@
 library(gstat)
 ptm <- proc.time()
-tracts <- readOGR("C:/Users/morrisk/Downloads/USA_Tracts", "USA_Tracts")
+tracts <- readOGR("raw_data/USA_Tracts", "USA_Tracts")
 
 rainfall <- readRDS("temp/rainfall_processed_2000_2020.rds") %>%
   ungroup()
@@ -22,24 +22,26 @@ rainsp <- SpatialPointsDataFrame(rainsp, rainfall)
 hold <- tracts
 
 for(s in unique(filter(fips_codes, state_code < 57)$state_code)){
-  tracts <- subset(hold, STATE_FIPS == s)
-  print(s)
-  
-  r <- raster(tracts, ncols = 100, nrows = 100)
-  ##############################
-  
-  gs <- gstat(formula=z~1, locations=rainsp)
-  idw <- interpolate(r, gs)
-  
-  districts.reprojected <- spTransform(tracts, CRS(projection(idw)))
-  
-  extracted.values <- raster::extract(idw, districts.reprojected)
-  
-  avs <- sapply(extracted.values, mean)
-  
-  tracts@data$z <- avs
-  saveRDS(tracts@data, paste0("temp/tracts_interp_", s, ".rds"))
-  proc.time() - ptm
+  if(!(file.exists(paste0("temp/tracts_interp_", s, ".rds")))){
+    tracts <- subset(hold, STATE_FIPS == s)
+    print(s)
+    
+    r <- raster(tracts, ncols = 100, nrows = 100)
+    ##############################
+    
+    gs <- gstat(formula=z~1, locations=rainsp)
+    idw <- interpolate(r, gs)
+    
+    districts.reprojected <- spTransform(tracts, CRS(projection(idw)))
+    
+    extracted.values <- raster::extract(idw, districts.reprojected)
+    
+    avs <- sapply(extracted.values, mean)
+    
+    tracts@data$z <- avs
+    saveRDS(tracts@data, paste0("temp/tracts_interp_", s, ".rds"))
+    proc.time() - ptm 
+  }
 }
